+++ migrations/2020-03-07_state_and_opening_hours/up.sql
@@ -1,0 +1,4 @@
+ALTER TABLE events ADD COLUMN state TEXT;
+
+ALTER TABLE place_revision ADD COLUMN state TEXT;
+ALTER TABLE place_revision ADD COLUMN opening_hours TEXT;